recommender_system/requirements.txt
gradio==3.39.0
torch==1.11.0
torchvision==0.12.0
numpy==1.22.3
Pillow==9.0.1
requests==2.27.1
flask==2.3.2
annoy==1.16.3

